Separate category listing into it's own components
