+++ package.json
@@ -73,6 +73,8 @@
     "react": "^16.12.0",
     "react-beautiful-dnd": "^12.2.0",
     "react-codemirror2": "^6.0.0",
+    "react-dnd": "^10.0.2",
+    "react-dnd-html5-backend": "^10.0.2",
     "react-dom": "^16.12.0",
     "react-feather": "^2.0.3",
     "react-helmet-async": "^1.0.4",
