+++ src/client/containers/AppSidebar.tsx
@@ -1,41 +1,19 @@
-import uuid from 'uuid/v4'
-import React, { useState, useRef, useEffect } from 'react'
-import {
-  Loader,
-  Folder as FolderIcon,
-  Plus,
-  Settings,
-  RefreshCw,
-  MoreHorizontal,
-} from 'react-feather'
+import React from 'react'
+import { Loader, Plus, Settings, RefreshCw } from 'react-feather'
 import { useDispatch, useSelector } from 'react-redux'
-import { Droppable, Draggable } from 'react-beautiful-dnd'
 
 import { LabelText } from '@resources/LabelText'
 import { TestID } from '@resources/TestID'
 
 import { ActionButton } from '@/components/AppSidebar/ActionButton'
-import { AddCategoryButton } from '@/components/AppSidebar/AddCategoryButton'
-import { AddCategoryForm } from '@/components/AppSidebar/AddCategoryForm'
 import { LastSyncedNotification } from '@/components/AppSidebar/LastSyncedNotification'
 import { AllNotesOption } from '@/components/AppSidebar/AllNotesOption'
 import { FolderOption } from '@/components/AppSidebar/FolderOption'
 import { ScratchpadOption } from '@/components/AppSidebar/ScratchpadOption'
-import { Folder, ContextMenuEnum } from '@/utils/enums'
-import { iconColor } from '@/utils/constants'
-import { useTempState } from '@/contexts/TempStateContext'
-import { ContextMenu } from '@/containers/ContextMenu'
+import { Folder } from '@/utils/enums'
+import { CategoryList } from '@/containers/CategoryList'
 import {
-  addCategory,
-  categoryDragEnter,
-  categoryDragLeave,
-  updateCategory,
-  setCategoryEdit,
-} from '@/slices/category'
-import {
-  addCategoryToNote,
   addNote,
-  updateActiveCategoryId,
   swapFolder,
   updateActiveNote,
   addFavoriteNote,
@@ -45,23 +23,15 @@ import {
 import { toggleSettingsModal, togglePreviewMarkdown } from '@/slices/settings'
 import { syncState } from '@/slices/sync'
 import { getSettings, getNotes, getCategories, getSync } from '@/selectors'
-import { CategoryItem, NoteItem, ReactDragEvent, ReactSubmitEvent, ReactMouseEvent } from '@/types'
-import {
-  newNoteHandlerHelper,
-  shouldOpenContextMenu,
-  determineCategoryClass,
-  getActiveNote,
-} from '@/utils/helpers'
+import { CategoryItem, NoteItem } from '@/types'
+import { newNoteHandlerHelper, getActiveNote } from '@/utils/helpers'
 
 export const AppSidebar: React.FC = () => {
   // ===========================================================================
   // Selectors
   // ===========================================================================
 
-  const {
-    categories,
-    editingCategory: { id: editingCategoryId, tempName: tempCategoryName },
-  } = useSelector(getCategories)
+  const { categories } = useSelector(getCategories)
   const { activeCategoryId, activeFolder, activeNoteId, notes } = useSelector(getNotes)
   const { previewMarkdown } = useSelector(getSettings)
   const { syncing, lastSynced } = useSelector(getSync)
@@ -79,39 +49,13 @@ export const AppSidebar: React.FC = () => {
     dispatch(updateActiveNote({ noteId, multiSelect }))
   const _updateSelectedNotes = (noteId: string, multiSelect: boolean) =>
     dispatch(updateSelectedNotes({ noteId, multiSelect }))
-  const _updateActiveCategoryId = (categoryId: string) =>
-    dispatch(updateActiveCategoryId(categoryId))
   const _swapFolder = (folder: Folder) => dispatch(swapFolder(folder))
-  const _addCategory = (category: CategoryItem) => dispatch(addCategory(category))
-  const _categoryDragEnter = (category: CategoryItem) => dispatch(categoryDragEnter(category))
-  const _categoryDragLeave = (category: CategoryItem) => dispatch(categoryDragLeave(category))
-  const _updateCategory = (category: CategoryItem) => dispatch(updateCategory(category))
   const _syncState = (notes: NoteItem[], categories: CategoryItem[]) =>
     dispatch(syncState({ notes, categories }))
   const _toggleSettingsModal = () => dispatch(toggleSettingsModal())
   const _togglePreviewMarkdown = () => dispatch(togglePreviewMarkdown())
   const _addTrashedNote = (noteId: string) => dispatch(addTrashedNote(noteId))
   const _addFavoriteNote = (noteId: string) => dispatch(addFavoriteNote(noteId))
-  const _addCategoryToNote = (categoryId: string, noteId: string) =>
-    dispatch(addCategoryToNote({ categoryId, noteId }))
-  const _setCategoryEdit = (categoryId: string, tempName: string) =>
-    dispatch(setCategoryEdit({ id: categoryId, tempName }))
-
-  // ===========================================================================
-  // Refs
-  // ===========================================================================
-  const contextMenuRef = useRef<HTMLDivElement>(null)
-
-  // ===========================================================================
-  // State
-  // ===========================================================================
-  const [optionsId, setOptionsId] = useState('')
-  const [optionsPosition, setOptionsPosition] = useState({ x: 0, y: 0 })
-
-  // ===========================================================================
-  // Context
-  // ===========================================================================
-  const { addingTempCategory, setAddingTempCategory } = useTempState()
 
   // ===========================================================================
   // Handlers
@@ -130,82 +74,8 @@ export const AppSidebar: React.FC = () => {
       _updateSelectedNotes
     )
 
-  const resetTempCategory = () => {
-    setAddingTempCategory(false)
-    _setCategoryEdit('', '')
-  }
-
-  const onSubmitNewCategory = (event: ReactSubmitEvent): void => {
-    event.preventDefault()
-
-    const category = { id: uuid(), name: tempCategoryName.trim(), draggedOver: false }
-
-    if (categories.find(cat => cat.name === category.name) || category.name === '') {
-      resetTempCategory()
-    } else {
-      _addCategory(category)
-      resetTempCategory()
-    }
-  }
-
-  const onSubmitUpdateCategory = (event: ReactSubmitEvent): void => {
-    event.preventDefault()
-
-    const category = { id: editingCategoryId, name: tempCategoryName.trim(), draggedOver: false }
-
-    if (categories.find(cat => cat.name === category.name) || category.name === '') {
-      resetTempCategory()
-    } else {
-      _updateCategory(category)
-      resetTempCategory()
-    }
-  }
-
-  const handleCategoryMenuClick = (
-    event: React.MouseEvent<HTMLDivElement, MouseEvent> | ReactMouseEvent,
-    categoryId: string = ''
-  ) => {
-    const clicked = event.target
-
-    // Make sure we aren't getting any null values .. any element clicked should be a sub-class of element
-    if (!clicked) return
-
-    if (shouldOpenContextMenu(clicked as Element)) {
-      if ('clientX' in event && 'clientY' in event) {
-        setOptionsPosition({ x: event.clientX, y: event.clientY })
-      }
-    }
-
-    event.stopPropagation()
-
-    if (contextMenuRef?.current?.contains(clicked as HTMLDivElement)) {
-      return
-    } else {
-      setOptionsId(!optionsId || optionsId !== categoryId ? categoryId : '')
-    }
-  }
-
-  const handleCategoryRightClick = (
-    event: React.MouseEvent<HTMLDivElement, MouseEvent> | ReactMouseEvent,
-    categoryId: string = ''
-  ) => {
-    event.preventDefault()
-    handleCategoryMenuClick(event, categoryId)
-  }
-
   const syncNotesHandler = () => _syncState(notes, categories)
   const settingsHandler = () => _toggleSettingsModal()
-
-  // ===========================================================================
-  // Hooks
-  // ===========================================================================
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleCategoryMenuClick)
-    return () => {
-      document.removeEventListener('mousedown', handleCategoryMenuClick)
-    }
-  })
 
   return (
     <>
@@ -252,129 +122,7 @@ export const AppSidebar: React.FC = () => {
           <div className="category-title">
             <h2>Categories</h2>
           </div>
-          <Droppable type="CATEGORY" droppableId="Category list">
-            {droppableProvided => (
-              <div
-                {...droppableProvided.droppableProps}
-                ref={droppableProvided.innerRef}
-                className="category-list"
-                aria-label="Category list"
-              >
-                {categories.map((category, index) => {
-                  return (
-                    <Draggable key={category.id} draggableId={category.id} index={index}>
-                      {(draggableProvided, snapshot) => (
-                        <div
-                          {...draggableProvided.dragHandleProps}
-                          {...draggableProvided.draggableProps}
-                          ref={draggableProvided.innerRef}
-                          data-testid={TestID.CATEGORY_LIST_DIV}
-                          className={determineCategoryClass(
-                            category,
-                            snapshot.isDragging,
-                            activeCategoryId
-                          )}
-                          onClick={() => {
-                            const notesForNewCategory = notes.filter(
-                              note => !note.trash && note.category === category.id
-                            )
-                            const newNoteId =
-                              notesForNewCategory.length > 0 ? notesForNewCategory[0].id : ''
-                            if (category.id !== activeCategoryId) {
-                              _updateActiveCategoryId(category.id)
-                              _updateActiveNote(newNoteId, false)
-                            }
-                          }}
-                          onDoubleClick={() => {
-                            _setCategoryEdit(category.id, category.name)
-                          }}
-                          onBlur={() => {
-                            _setCategoryEdit('', '')
-                          }}
-                          onDrop={event => {
-                            event.preventDefault()
-
-                            _addCategoryToNote(category.id, event.dataTransfer.getData('text'))
-                            _categoryDragLeave(category)
-                          }}
-                          onDragOver={(event: ReactDragEvent) => event.preventDefault()}
-                          onDragEnter={() => _categoryDragEnter(category)}
-                          onDragLeave={() => _categoryDragLeave(category)}
-                          onContextMenu={event => handleCategoryRightClick(event, category.id)}
-                        >
-                          <form
-                            className="category-list-name"
-                            onSubmit={event => {
-                              event.preventDefault()
-                              _setCategoryEdit('', '')
-                              onSubmitUpdateCategory(event)
-
-                              if (optionsId) setOptionsId('')
-                            }}
-                          >
-                            <FolderIcon size={15} className="app-sidebar-icon" color={iconColor} />
-                            {editingCategoryId === category.id ? (
-                              <input
-                                data-testid={TestID.CATEGORY_EDIT}
-                                className="category-edit"
-                                type="text"
-                                autoFocus
-                                maxLength={20}
-                                value={tempCategoryName}
-                                onChange={event => {
-                                  _setCategoryEdit(editingCategoryId, event.target.value)
-                                }}
-                                onBlur={event => onSubmitUpdateCategory(event)}
-                              />
-                            ) : (
-                              category.name
-                            )}
-                          </form>
-                          <div
-                            data-testid={TestID.MOVE_CATEGORY}
-                            className={
-                              optionsId === category.id
-                                ? 'category-options active'
-                                : 'category-options'
-                            }
-                            onClick={event => handleCategoryMenuClick(event, category.id)}
-                          >
-                            <MoreHorizontal size={15} className="context-menu-action" />
-                          </div>
-                          {optionsId === category.id && (
-                            <ContextMenu
-                              contextMenuRef={contextMenuRef}
-                              item={category}
-                              optionsPosition={optionsPosition}
-                              setOptionsId={setOptionsId}
-                              type={ContextMenuEnum.CATEGORY}
-                            />
-                          )}
-                        </div>
-                      )}
-                    </Draggable>
-                  )
-                })}
-                {droppableProvided.placeholder}
-              </div>
-            )}
-          </Droppable>
-          {addingTempCategory ? (
-            <AddCategoryForm
-              dataTestID={TestID.NEW_CATEGORY_FORM}
-              submitHandler={onSubmitNewCategory}
-              changeHandler={_setCategoryEdit}
-              resetHandler={resetTempCategory}
-              editingCategoryId={editingCategoryId}
-              tempCategoryName={tempCategoryName}
-            />
-          ) : (
-            <AddCategoryButton
-              dataTestID={TestID.ADD_CATEGORY_BUTTON}
-              handler={setAddingTempCategory}
-              label={LabelText.ADD_CATEGORY}
-            />
-          )}
+          <CategoryList />
         </section>
       </aside>
       {lastSynced && <LastSyncedNotification datetime={lastSynced} />}
